fix nm invocation on OpenBSD
Apparently it's a bad idea to close stdin and then execvp() something.
I don't know what the exact mechanism was, but nm(1) on OpenBSD would
tolerate this when invoked as 'nm -P file'.  When invoked as 'nm -CP
file' however, nm(1) itself forks and invokes c++filt, and in that
scenario it gave no output.

Giving it /dev/null as stdin seems to resolve the issue.